# swg_api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DADS API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://portal.dratdoorstep.com/api"
        },
        {
            "url": "https://deviboon.dratdoorstep.com/api"
        },
        {
            "url": "http://localhost/DADS_Web/api"
        }
    ],
    "paths": {
...
        "/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Blog",
                "description": "",
                "operationId": "getBlog",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid/unauthorized token or token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Api_Response"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "dads_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Api_Response": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "dads_auth": {
                "type": "http",
                "name": "dads_auth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}